requirements.txt
openai>=1.82.0
pyyaml>=6.0
python-dotenv>=1.1.0
